Justert trend data til trend linje funksjon (TredLine) Endret lengde på trendlinjer fra 300 til 340.

+++ HMS_Client/Helpers/TrendLine.cs
@@ -11,7 +11,7 @@ namespace HMS_Client
     {
         public static void GenerateGridColumnDefinitions(Canvas canvas, double listMaxLength)
         {
-            int width = (int)(canvas.Width / listMaxLength);
+            double width = (int)(canvas.Width / listMaxLength);
 
             for (int i = 0; i < listMaxLength; i++)
             {
